# cli/openapi/agent-openapi.json
{
  "components": {
    "schemas": {
...
      "AgentDigestRequest": {
        "description": "Agent digest generation request for arbitrary windows.",
        "properties": {
          "end_at": {
            "format": "date-time",
            "title": "End At",
            "type": "string"
          },
          "form": {
            "default": "short",
            "enum": [
              "short",
              "long"
            ],
            "title": "Form",
            "type": "string"
          },
          "start_at": {
            "format": "date-time",
            "title": "Start At",
            "type": "string"
          }
        },
        "required": [
          "start_at",
          "end_at"
        ],
        "title": "AgentDigestRequest",
        "type": "object"
      },
      "AgentDigestResponse": {
        "description": "Async agent digest generation response.",
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id"
        ],
        "title": "AgentDigestResponse",
        "type": "object"
      },
...
      "AgentLibraryFileResponse": {
        "description": "One markdown document payload for CLI sync download.",
        "properties": {
          "checksum_sha256": {
            "title": "Checksum Sha256",
            "type": "string"
          },
          "content_id": {
            "title": "Content Id",
            "type": "integer"
          },
          "relative_path": {
            "title": "Relative Path",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "nullable": true,
            "title": "Updated At",
            "type": "string"
          },
          "variant": {
            "enum": [
              "source",
              "summary"
            ],
            "title": "Variant",
            "type": "string"
          }
        },
        "required": [
          "relative_path",
          "content_id",
          "variant",
          "checksum_sha256",
          "text"
        ],
        "title": "AgentLibraryFileResponse",
        "type": "object"
      },
      "AgentLibraryManifestResponse": {
        "description": "Manifest of markdown documents available for CLI sync.",
        "properties": {
          "documents": {
            "items": {
              "$ref": "#/components/schemas/AgentLibraryDocumentResponse"
            },
            "title": "Documents",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "include_source": {
            "default": true,
            "title": "Include Source",
            "type": "boolean"
          }
        },
        "required": [
          "generated_at",
          "documents"
        ],
        "title": "AgentLibraryManifestResponse",
        "type": "object"
      },
      "AgentOnboardingCompleteRequest": {
        "description": "Complete simplified agent onboarding.",
        "properties": {
          "accept_all": {
            "default": false,
            "title": "Accept All",
            "type": "boolean"
          },
          "selected_aggregators": {
            "items": {
              "$ref": "#/components/schemas/OnboardingSelectedAggregator"
            },
            "title": "Selected Aggregators",
            "type": "array"
          },
          "selected_subreddits": {
            "items": {
              "type": "string"
            },
            "title": "Selected Subreddits",
            "type": "array"
          },
          "source_ids": {
            "items": {
              "type": "integer"
            },
            "title": "Source Ids",
            "type": "array"
          }
        },
        "title": "AgentOnboardingCompleteRequest",
        "type": "object"
      },
      "AgentOnboardingStartRequest": {
        "description": "Simplified async onboarding start request.",
        "properties": {
          "brief": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Brief",
            "type": "string"
          },
          "preferences": {
            "additionalProperties": true,
            "nullable": true,
            "title": "Preferences",
            "type": "object"
          },
          "seed_feeds": {
            "items": {
              "type": "string"
            },
            "title": "Seed Feeds",
            "type": "array"
          },
          "seed_urls": {
            "items": {
              "type": "string"
            },
            "title": "Seed Urls",
            "type": "array"
          }
        },
        "required": [
          "brief"
        ],
        "title": "AgentOnboardingStartRequest",
        "type": "object"
      },
      "AgentOnboardingStartResponse": {
        "description": "Simplified async onboarding start response.",
        "properties": {
          "job_id": {
            "nullable": true,
            "title": "Job Id",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "AgentOnboardingStartResponse",
        "type": "object"
      },
      "AgentSearchRequest": {
        "description": "Machine-oriented external search request.",
        "properties": {
          "include_podcasts": {
            "default": true,
            "title": "Include Podcasts",
            "type": "boolean"
          },
          "limit": {
            "default": 10,
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "query": {
            "maxLength": 200,
            "minLength": 2,
            "title": "Query",
            "type": "string"
          }
        },
        "required": [
          "query"
        ],
        "title": "AgentSearchRequest",
        "type": "object"
      },
...
      "CliLinkApproveRequest": {
        "description": "Authenticated request to approve one pending CLI link session.",
        "properties": {
          "approve_token": {
            "maxLength": 255,
            "minLength": 8,
            "title": "Approve Token",
            "type": "string"
          },
          "device_name": {
            "maxLength": 255,
            "nullable": true,
            "title": "Device Name",
            "type": "string"
          }
        },
        "required": [
          "approve_token"
        ],
        "title": "CliLinkApproveRequest",
        "type": "object"
      },
      "CliLinkApproveResponse": {
        "description": "Approval response after issuing a CLI API key.",
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "key_prefix": {
            "title": "Key Prefix",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "const": "approved",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "key_prefix",
          "expires_at"
        ],
        "title": "CliLinkApproveResponse",
        "type": "object"
      },
      "CliLinkPollResponse": {
        "description": "Polling response for a CLI waiting on mobile approval.",
        "properties": {
          "api_key": {
            "nullable": true,
            "title": "Api Key",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "key_prefix": {
            "nullable": true,
            "title": "Key Prefix",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "enum": [
              "pending",
              "approved",
              "claimed",
              "expired"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "expires_at"
        ],
        "title": "CliLinkPollResponse",
        "type": "object"
      },
      "CliLinkStartRequest": {
        "description": "Request to start a new CLI QR-link session.",
        "properties": {
          "device_name": {
            "maxLength": 255,
            "nullable": true,
            "title": "Device Name",
            "type": "string"
          }
        },
        "title": "CliLinkStartRequest",
        "type": "object"
      },
      "CliLinkStartResponse": {
        "description": "Unauthenticated response for bootstrapping CLI QR login.",
        "properties": {
          "approve_url": {
            "title": "Approve Url",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "poll_interval_seconds": {
            "default": 2,
            "title": "Poll Interval Seconds",
            "type": "integer"
          },
          "poll_token": {
            "title": "Poll Token",
            "type": "string"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "status": {
            "const": "pending",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "session_id",
          "status",
          "poll_token",
          "approve_url",
          "expires_at"
        ],
        "title": "CliLinkStartResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OnboardingDiscoveryStatusResponse": {
        "description": "Status response for onboarding audio discovery polling.",
        "properties": {
          "error_message": {
            "nullable": true,
            "title": "Error Message",
            "type": "string"
          },
          "inferred_topics": {
            "items": {
              "type": "string"
            },
            "title": "Inferred Topics",
            "type": "array"
          },
          "lanes": {
            "items": {
              "$ref": "#/components/schemas/OnboardingDiscoveryLaneStatus"
            },
            "title": "Lanes",
            "type": "array"
          },
          "run_id": {
            "title": "Run Id",
            "type": "integer"
          },
          "run_status": {
            "title": "Run Status",
            "type": "string"
          },
          "suggestions": {
            "$ref": "#/components/schemas/OnboardingFastDiscoverResponse",
            "nullable": true
          },
          "topic_summary": {
            "nullable": true,
            "title": "Topic Summary",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "run_status"
        ],
        "title": "OnboardingDiscoveryStatusResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Filtered machine-oriented API contract for the standalone newsly-agent CLI.",
    "title": "Newsly Agent CLI API",
    "version": "2.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
    "/api/agent/cli/link/start": {
      "post": {
        "description": "Create an unauthenticated QR approval session for the CLI.",
        "operationId": "startCliLink",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CliLinkStartRequest",
                "nullable": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliLinkStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Cli Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/agent/cli/link/{session_id}": {
      "get": {
        "description": "Poll a pending CLI link session without requiring existing auth.",
        "operationId": "pollCliLink",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "poll_token",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 8,
              "title": "Poll Token",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliLinkPollResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Poll Cli Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/agent/cli/link/{session_id}/approve": {
      "post": {
        "description": "Approve one pending CLI link session from the authenticated app.",
        "operationId": "approveCliLink",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CliLinkApproveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CliLinkApproveResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Approve Cli Link",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/agent/digests": {
      "post": {
        "description": "Queue arbitrary-window digest generation for agent clients.",
        "operationId": "generateDigest",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDigestRequest",
                "nullable": true,
                "title": "Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDigestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Generate Digest",
        "tags": [
          "news"
        ]
      }
    },
    "/api/agent/library/file": {
      "get": {
        "description": "Return one rendered markdown document by relative manifest path.",
        "operationId": "getAgentLibraryFile",
        "parameters": [
          {
            "in": "query",
            "name": "path",
            "required": true,
            "schema": {
              "maxLength": 1024,
              "minLength": 1,
              "title": "Path",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentLibraryFileResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Library File",
        "tags": [
          "library"
        ]
      }
    },
    "/api/agent/library/manifest": {
      "get": {
        "description": "Return manifest metadata for exportable per-user markdown files.",
        "operationId": "getAgentLibraryManifest",
        "parameters": [
          {
            "in": "query",
            "name": "include_source",
            "required": false,
            "schema": {
              "default": true,
              "title": "Include Source",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentLibraryManifestResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Agent Library Manifest",
        "tags": [
          "library"
        ]
      }
    },
    "/api/agent/onboarding": {
      "post": {
        "description": "Start simplified async onboarding.",
        "operationId": "startOnboarding",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentOnboardingStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentOnboardingStartResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Start Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/agent/onboarding/{run_id}": {
      "get": {
        "description": "Return onboarding run status.",
        "operationId": "getOnboarding",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OnboardingDiscoveryStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/agent/onboarding/{run_id}/complete": {
      "post": {
        "description": "Complete onboarding from simplified selections.",
        "operationId": "completeOnboarding",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentOnboardingCompleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Completeonboarding",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Complete Onboarding",
        "tags": [
          "onboarding"
        ]
      }
    },
    "/api/agent/search": {
      "post": {
        "description": "Search external/provider-backed sources for the agent CLI.",
        "operationId": "searchAgent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Agent",
        "tags": [
          "search"
        ]
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "description": "Return async job status.",
        "operationId": "getJob",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
...
  },
...
}